The robosim validator inherits the checks from robochart, but in those cases it uses the robochart type checker. for this reason, elements like
$v = $x

yield an error, because robochart cannot get the type of $x.

In order to fix this, I had to override the assignment check in the
validator of robosim (with the same behaviour as in robochart).

I need to figure out a way of parameterising the validator with the
correct type checker.